% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERA5.R
\docType{data}
\name{ERA5}
\alias{ERA5}
\title{ERA5 Autumn 3-day precipitation extracted for the West-coast of Norway}
\format{
An object of class `"data.frame"`
}
\usage{
read.csv("Data/ERA5.csv")
}
\description{
ERA5 Autumn 3-day precipitation extracted for the West-coast of Norway
}
\references{
Kelder et al. 2020
}
\keyword{datasets}
