% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fidelity.R, R/Independence.R, R/evaluate.r
\name{Independence_test}
\alias{Independence_test}
\title{Calculate the correlation between each unique pair of ensemble members.}
\usage{
Independence_test(ensemble, n_ensembles = 25, n_lds = 5)

Independence_test(ensemble, n_ensembles = 25, n_lds = 5)

Independence_test(ensemble, n_ensembles = 25, n_lds = 5)
}
\arguments{
\item{ensemble}{The UNSEEN ensemble. This function expects an dataframe with variables ensemble, leadtime, precipitation.
It expects that the first member is 0 (as used in CDS) and that the first lead time is 2 
(first month shows dependence and is therefore removed).}
}
\value{
a plot showing the boxplot of the pairwise correlations for each leadtime. See Kelder et al. 2020

a plot showing the boxplot of the pairwise correlations for each leadtime. See Kelder et al. 2020

a plot showing the boxplot of the pairwise correlations for each leadtime. See Kelder et al. 2020
}
\description{
Calculate the correlation between each unique pair of ensemble members.

Calculate the correlation between each unique pair of ensemble members.

Calculate the correlation between each unique pair of ensemble members.
}
\section{Warning}{

The confidence intervals are based on a bootstrap test with n_ensembles = 25. This will not give the right CI bounds for different
ensemble sizes.


The confidence intervals are based on a bootstrap test with n_ensembles = 25. This will not give the right CI bounds for different
ensemble sizes.


The confidence intervals are based on a bootstrap test with n_ensembles = 25. This will not give the right CI bounds for different
ensemble sizes.
}

