% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fidelity.R, R/Stability.R, R/evaluate.r
\name{Model_stability_boot}
\alias{Model_stability_boot}
\title{Test the model stability: 2. empirical extreme value plot}
\source{
Evaluation explaned in more detail in Kelder et al. 2020

Colorblind friendly palette  http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/#a-colorblind-friendly-palette

Evaluation explaned in more detail in Kelder et al. 2020

Colorblind friendly palette  http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/#a-colorblind-friendly-palette

Evaluation explaned in more detail in Kelder et al. 2020

Colorblind friendly palette  http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/#a-colorblind-friendly-palette
}
\usage{
Model_stability_boot(ensemble, lab = "")

Model_stability_boot(ensemble, lab = "")

Model_stability_boot(ensemble, lab = "")
}
\arguments{
\item{ensemble}{The UNSEEN ensemble. This function expects an dataframe with variables leadtime, precipitation.}
}
\value{
a plot with the empirical return values of the pooled ensemble including confidence intervals. 
Individual lead times are plotted on top.

a plot with the empirical return values of the pooled ensemble including confidence intervals. 
Individual lead times are plotted on top.

a plot with the empirical return values of the pooled ensemble including confidence intervals. 
Individual lead times are plotted on top.
}
\description{
We plot the density distribution of the different leadtimes using ggplot.   
We want to show the confidence interval of the distribution of all lead times pooled together,
and test whether the individual lead times fall within these confidence intervals. 
Therefore we bootstrap the pooled leadtimes into series with equal length of the individual leadtimes (875), with n=10000.

We plot the density distribution of the different leadtimes using ggplot.   
We want to show the confidence interval of the distribution of all lead times pooled together,
and test whether the individual lead times fall within these confidence intervals. 
Therefore we bootstrap the pooled leadtimes into series with equal length of the individual leadtimes (875), with n=10000.

We plot the density distribution of the different leadtimes using ggplot.   
We want to show the confidence interval of the distribution of all lead times pooled together,
and test whether the individual lead times fall within these confidence intervals. 
Therefore we bootstrap the pooled leadtimes into series with equal length of the individual leadtimes (875), with n=10000.
}
