% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fidelity.R, R/Stability.R, R/evaluate.r
\name{Model_stability_density}
\alias{Model_stability_density}
\title{Test the model stability: 1. density plot}
\source{
Evaluation explaned in more detail in Kelder et al. 2020

Colorblind friendly palette  http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/#a-colorblind-friendly-palette

Evaluation explaned in more detail in Kelder et al. 2020

Colorblind friendly palette  http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/#a-colorblind-friendly-palette

Evaluation explaned in more detail in Kelder et al. 2020

Colorblind friendly palette  http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/#a-colorblind-friendly-palette
}
\usage{
Model_stability_density(ensemble, lab = "")

Model_stability_density(ensemble, lab = "")

Model_stability_density(ensemble, lab = "")
}
\arguments{
\item{ensemble}{The UNSEEN ensemble. This function expects an dataframe with variables leadtime, precipitation.}
}
\value{
a plot showing the empirical probability density distribution for each leadtime

a plot showing the empirical probability density distribution for each leadtime

a plot showing the empirical probability density distribution for each leadtime
}
\description{
We plot the density distribution of the different leadtimes using ggplot.

We plot the density distribution of the different leadtimes using ggplot.

We plot the density distribution of the different leadtimes using ggplot.
}
