% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fidelity.R, R/evaluate.r
\name{plot_unseen}
\alias{plot_unseen}
\title{Illustrate the timeseries of the UNsEEN ensemble and the observed using ggplot}
\usage{
plot_unseen(ensemble, obs, ylab = "", title = "")

plot_unseen(ensemble, obs, ylab = "", title = "")
}
\arguments{
\item{ensemble}{The UNSEEN ensemble. This must be a dataframe with at least the variables year and precipitation}

\item{obs}{The observations. This must be a dataframe with the variables year and precipitation}

\item{ylab}{The y label (string)}

\item{title}{The title (string)}
}
\value{
a ggplot with x = year and y =

a ggplot with x = year and y =
}
\description{
Illustrate the timeseries of the UNsEEN ensemble and the observed using ggplot

Illustrate the timeseries of the UNsEEN ensemble and the observed using ggplot
}
