% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAS5_wc.R
\docType{data}
\name{SEAS5_wc}
\alias{SEAS5_wc}
\title{SEAS5 UNSEEN ensemble extracted for the West-coast of Norway}
\format{
An object of class `"data.frame"`
}
\usage{
read.csv("Data/SEAS5_wc.rda")
}
\description{
SEAS5 UNSEEN ensemble extracted for the West-coast of Norway
}
\references{
Kelder et al. 2020
}
\keyword{datasets}
