% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeNorge.R
\docType{data}
\name{SeNorge}
\alias{SeNorge}
\title{SeNorge Autumn 3-day precipitation extremes extracted for the West-coast of Norway}
\format{
An object of class `"data.frame"`
}
\usage{
read.csv("Data/SeNorge.csv")
}
\description{
SeNorge Autumn 3-day precipitation extremes extracted for the West-coast of Norway
}
\references{
Kelder et al. 2020
}
\keyword{datasets}
