% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_plot.R
\name{unseen_timeseries}
\alias{unseen_timeseries}
\title{Illustrate the timeseries of the UNsEEN ensemble and the observed using ggplot}
\usage{
unseen_timeseries(ensemble, obs, ylab = "", title = "")
}
\arguments{
\item{ensemble}{The UNSEEN ensemble. This must be a dataframe with at least the variables year and precipitation}

\item{obs}{The observations. This must be a dataframe with the variables year and precipitation}

\item{ylab}{The y label (string)}

\item{title}{The title (string)}
}
\value{
a ggplot timeseries with x = year and y = variable
}
\description{
Illustrate the timeseries of the UNsEEN ensemble and the observed using ggplot
}
